Adding some comments in terms of what I would do
Since this isn't a gist, I figured this could also be a good way to give you some feedback & actually like that this allows me to inline comment
My comments are marked with CJ, but you can also see it in the diff

Logic is there besides returning a value from the method. My comments are largely 'optional' code hygiene. Good job!